refactor: add ProductHelperService
