+++ src/app/features/product/product-list/product-list.component.ts
@@ -4,8 +4,9 @@ import { Router } from '@angular/router';
 
 import { BriefCardComponent } from './brief-card/brief-card.component';
 import { Product } from '../../../core/product/interfaces/product';
-import { ProductService } from '../../../core/product/product.service';
 import { ProductResponse } from '../../../core/product/interfaces/product-response';
+import { ProductService } from '../../../core/product/product.service';
+import { ProductHelperService } from '../../../core/product/product.helper.service';
 import { NavigateToSpecificRouteService } from '../../../shared/services/navigate-to-specific-route/navigate-to-specific-route.service';
 
 @Component({
@@ -17,7 +18,8 @@ import { NavigateToSpecificRouteService } from '../../../shared/services/navigat
 export class ProductListComponent implements OnInit {
   constructor(
     private navigateToSpecificRouteService: NavigateToSpecificRouteService,
-    private productService: ProductService,
+    private apiService: ProductService,
+    public helperService: ProductHelperService,
     private router: Router,
   ) {}
 
@@ -28,7 +30,7 @@ export class ProductListComponent implements OnInit {
   public ngOnInit() {
     this.currentRoute = this.router.url;
 
-    this.productService.getAllProducts().subscribe({
+    this.apiService.getAllProducts().subscribe({
       next: (response) => {
         const responseStr = JSON.stringify(response);
         const productResponse: ProductResponse = JSON.parse(responseStr);
@@ -52,27 +54,6 @@ export class ProductListComponent implements OnInit {
     }
 
     return new Error('Something went wrong; please try again later.');
-  }
-
-  public getProductImg(product: Product): string {
-    const currentMasterVariantImage = product.masterData.current.masterVariant.images[0];
-
-    return currentMasterVariantImage.url;
-  }
-
-  public getProductName(product: Product): string {
-    const currentName = product.masterData.current.name;
-    const key = Object.keys(currentName)[0];
-
-    return currentName[key];
-  }
-
-  public getShortProductDescription(product: Product): string {
-    const currentMetaDescription = product.masterData.current.metaDescription;
-    const key = Object.keys(currentMetaDescription)[0];
-    const description = currentMetaDescription[key];
-
-    return description.length > 89 ? `${description.slice(0, 80)}...` : description;
   }
 
   public goBack(): void {
